pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-kit"
version = "0.1.0"
description = "Reusable prompt fragments, fixtures, and evaluation harnesses"
authors = [{name = ""}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = []
dynamic = []

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pyyaml>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["prompt_kit"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
